requirements.txt
altgraph==0.17.3
attrs==23.1.0
Cartopy==0.22.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.0
cycler==0.11.0
ephem==4.1.4
et-xmlfile==1.1.0
Fiona==1.9.4.post1
fonttools==4.42.1
geographiclib==2.0
geopandas==0.13.2
geopy==2.4.0
idna==3.4
jplephem==2.18
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.0.9
pyproj==3.6.0
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scipy==1.11.2
screeninfo==0.8.1
sgp4==2.22
shapely==2.0.1
six==1.16.0
skyfield==1.46
suncalc==0.1.3
tzdata==2023.3
urllib3==2.0.4
